+++ src/routing.rs
@@ -1,12 +1,15 @@
+pub mod api;
 pub mod basic;
+pub mod downs;
+pub mod froggi_middleware;
 pub mod login;
+pub mod misc;
 pub mod overlay;
+pub mod popups;
 pub mod sponsors;
 pub mod team;
 pub mod teaminfo;
 pub mod time;
+pub mod updating;
 pub mod visibility;
 pub mod websockets;
-pub mod downs;
-pub mod api;
-pub mod updating;
