+++ src/routing/api.rs
@@ -1,6 +1,10 @@
 use std::collections::HashMap;
 
-use axum::{extract::{Path, State}, http::{HeaderMap, HeaderName, HeaderValue, StatusCode}, response::{Html, IntoResponse, Response}};
+use axum::{
+    extract::{Path, State},
+    http::{HeaderMap, HeaderName, HeaderValue, StatusCode},
+    response::{Html, IntoResponse, Response},
+};
 
 use crate::{appstate::global::*, key_create, printlg, AppState, Login, API_KEY_LEN};
