+++ src/routing/popups.rs
@@ -1,0 +1,70 @@
+use axum::{
+    extract::{Path, Query},
+    http::StatusCode,
+    response::IntoResponse,
+};
+use std::{collections::HashMap, time::Instant};
+
+use crate::{appstate::global::*, printlg};
+
+pub async fn popup_handler(
+    Path(a): Path<String>,
+    Query(params): Query<HashMap<String, String>>,
+) -> impl IntoResponse {
+    if let Some(p) = params.get("text") {
+        if a == "home" {
+            POPUPS_HOME.lock().await.push((p.clone(), 7));
+        } else if a == "away" {
+            POPUPS_AWAY.lock().await.push((p.clone(), 7));
+        }
+        printlg!("POPUP: {}", p);
+    }
+
+    return StatusCode::OK;
+}
+
+pub async fn popup_home_ticker() {
+    loop {
+        let start_time = Instant::now();
+        tokio::time::sleep(std::time::Duration::from_secs(1)).await;
+        let mut popups = POPUPS_HOME.lock().await;
+
+        let mut i = 0;
+        loop {
+            if i >= popups.len() {
+                break;
+            }
+
+            let time_diff = (Instant::now() - start_time).as_secs();
+            if popups[i].1 - time_diff > 0 {
+                popups[i].1 -= time_diff;
+                i += 1;
+            } else {
+                popups.remove(i);
+            }
+        }
+    }
+}
+
+pub async fn popup_away_ticker() {
+    loop {
+        let start_time = Instant::now();
+        tokio::time::sleep(std::time::Duration::from_secs(1)).await;
+        let mut popups = POPUPS_AWAY.lock().await;
+
+        let mut i = 0;
+        loop {
+            if i >= popups.len() {
+                break;
+            }
+
+            let time_diff = (Instant::now() - start_time).as_secs();
+            if popups[i].1 - time_diff > 0 {
+                popups[i].1 -= time_diff;
+                i += 1;
+            } else {
+                popups.remove(i);
+            }
+        }
+    }
+}
